amend image field for add/edit product

+++ products/views.py
@@ -1,13 +1,13 @@
 """ A view to return the products template """
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from .models import Product, Category
 from .forms import ProductForm
 
 
 def all_products(request):
     """ Return shop page and show all products """
-
     products = Product.objects.all()
     categories = None
 
@@ -35,8 +35,13 @@ def product_details(request, product_id):
     return render(request, 'products/product_details.html', context)
 
 
+@login_required
 def add_product(request):
     """ For adding products to the shop """
+    if not request.user.is_superuser:
+        messages.error(request, 'Only shop staff have access to this.')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -57,8 +62,13 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request, product_id):
     """ Edit product already in the shop """
+    if not request.user.is_superuser:
+        messages.error(request, 'Only shop staff have access to this.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -82,8 +92,13 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
     """ Delete product from the shop """
+    if not request.user.is_superuser:
+        messages.error(request, 'Only shop staff have access to this.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'Product deleted.')
